added BATAVIA_MAGIC which allows for Py3.4 and Py3.5 support
The deal has to do with that fact that byte code has slightly changed
in Python3.5. This change attempts to add support for batavia for both
Py3.4 and Py3.5 by storing the byte code magic number in a window
global hence allowing to check for it in the runloop.

The issue is related to https://github.com/pybee/batavia/issues/26
as well as https://github.com/pybee/batavia/issues/41.