Typo
Very minor typo.  On a related note, the code-block:: doscom doesn't seem to show up anywhere - not in the github version of these docs or the hosted readthedocs.  Similarly, the code block for anaconda doesn't appear (the page says For those using anaconda: and then there's nothing).  I don't quite know how to fix that