Basic frozenset implementation
This has revealed some bugs in the way we handle hash table things, like
dicts, sets, and frozensets. For example, we don't differentiate well
between str("1") and int(1) in keys. We'll have to handle this better,
by building and managing our own hash tables eventually, but this
will do for now.