fix failures exposed with new argument validation
isinstance now expects to receive arguments that are types. The normal type names are in the global scope as builtin functions, but not types. Prior to the update to the argument validation, isinstance returned False for things like isinstance(1, int), but true for isinstance(1, type(1)).

The next step here is to make `int == type(1)` and such for all the type names.