Add tests for calling magic methods directly.
I fixed some of the new tests, but in the process broke a few of the existing ones. They are still throwing the right type of exception, but the error message doesn't match exactly. There is a special case that throws the error "can't multiply sequence by a non-int of type'blah'" but it is slightly inconsistent in the case of inplace operators, so I've punted on that for now.